Move some config consts to a separate file
Created the file constants.js so that a human can do
configuration for values that will want to be changed all
in one place. Moved a few values, named them nicely,
and tested that changing them still works.

Watch out for future commits that move more configuration
constants to this file.